move persistence tests to own folder
