+++ jest.config.js
@@ -1,5 +1,4 @@
 module.exports = { 
   "preset": "jest-puppeteer", 
-  "testRegex": "(/__tests__/Test.*|(\\.|/)(test|spec))\\.[jt]sx?$",
-  "testPathIgnorePatterns" : ["__tests__/TestPersistence.js"]
+  "testRegex": "(/__tests__/Test.*|(\\.|/)(test|spec))\\.[jt]sx?$"
 };
